python/requirements.txt
feedparser==6.0.8
requests==2.26.0
python-dateutil==2.8.2
python-dotenv==0.19.2